refactor: extract chat pane into ui/chat_pane.rs (#357)
Fifteenth slice. The big one: draw_messages (~676 lines) plus the
two functions tightly coupled to it.

Moved to ui/chat_pane.rs:
- draw_chat_area              (layout dispatcher: messages + composer)
- draw_messages               (per-conversation rendering pipeline)
- patch_kitty_placeholders    (post-draw Kitty graphics image patch)
- emoji_to_text               (private helper)
- build_reaction_summary      (was pub(crate), now private)
- build_poll_display          (was pub(crate), now private)
- All related tests           (build_reaction_summary_*, build_poll_display_*)

Several mod.rs helpers became pub(super) so chat_pane can call them:
MSG_WINDOW_MULTIPLIER, collect_link_regions, split_spans_by_newline,
styled_uri_spans. The composer's draw_input is reached via super::composer,
welcome's draw_welcome via super::welcome.

ui/mod.rs: 2218 -> 1102 lines (79% total reduction from the original 5272).
